+++ stl/daydayup/tigerwe/src/task_1/main.cpp
@@ -1,17 +1,17 @@
 #include <iostream>
-#include "stl_list.h"
+#include "tigerwe_list.h"
 
 class Test {
  public:
-  Test() { std::cout << "construct.." << std::endl; }
+  Test() { std::cout << "list construct.." << std::endl; }
 
-  void method() { std::cout << "welcome Test.." << std::endl; }
+  void method() { std::cout << "Test.." << std::endl; }
 
-  ~Test() { std::cout << "destruct.." << std::endl; }
+  ~Test() { std::cout << "list destruct.." << std::endl; }
 };
 
-void printfList(Yami::list<int>& list_INT) {
-  Yami::list<int>::list_iterator itrBegin = list_INT.begin();
+void printfList(list<int>& list_INT) {
+  list<int>::iterator itrBegin = list_INT.begin();
 
   while (itrBegin != list_INT.end()) {
     std::cout << *itrBegin;
@@ -22,23 +22,27 @@ void printfList(Yami::list<int>& list_INT) {
 }
 
 int main(int argc, char* argv[]) {
-  std::cout << "Test bdgin !" << std::endl;
+  std::cout << "Test begin !" << std::endl;
   // test int
-  Yami::list<int> list_INT;
+  list<int> list_INT;
+  // 尾插
+  list_INT.push_back(3);
   list_INT.push_back(1);
   list_INT.push_back(2);
-  list_INT.push_back(3);
   list_INT.push_back(4);
   list_INT.push_back(5);
   list_INT.push_back(2);
 
   printfList(list_INT);
 
+  // 删除指定节点   这里返回的是删除的个数
+  std::cout << "delete nums: " << list_INT.remove(2) << std::endl;
   std::cout << "delete nums: " << list_INT.remove(2) << std::endl;
 
   printfList(list_INT);
 
-  Yami::list<int> list_INT1;
+  list<int> list_INT1;
+  // 头插
   list_INT1.push_front(1);
   list_INT1.push_front(2);
   list_INT1.push_front(3);
@@ -46,11 +50,12 @@ int main(int argc, char* argv[]) {
   list_INT1.push_front(5);
 
   printfList(list_INT1);
-
+  // 访问
   std::cout << "front: " << list_INT1.front() << std::endl;
 
   std::cout << "back: " << list_INT1.back() << std::endl;
 
+  // 删除
   list_INT1.pop_back();
 
   list_INT1.pop_front();
@@ -60,10 +65,11 @@ int main(int argc, char* argv[]) {
   printfList(list_INT1);
 
   // test class  主要看一下资源析构情况
+  // 这里每个会调用两次构造函数
   Test test1;
   Test test2;
   Test test3;
-  Yami::list<Test> list_CLASS;
+  list<Test> list_CLASS;
   list_CLASS.push_back(test1);
   list_CLASS.push_back(test2);
   list_CLASS.push_back(test3);
@@ -75,17 +81,17 @@ int main(int argc, char* argv[]) {
   std::cout << list_CLASS.size() << std::endl;
 
   // test string
-  Yami::list<std::string> list_STRING;
+  list<std::string> list_STRING;
 
-  list_STRING.push_back("nihao");
-  list_STRING.push_back("thanks");
-  list_STRING.push_back("goodbye");
-  list_STRING.push_back("seeyou");
+  list_STRING.push_back("vivwimp");
+  list_STRING.push_back("always");
+  list_STRING.push_back("miss");
+  list_STRING.push_back("babysong");
 
-  Yami::list<std::string>::list_iterator itBegin = list_STRING.begin();
+  list<std::string>::iterator itBegin = list_STRING.begin();
 
   while (itBegin != list_STRING.end()) {
-    std::cout << " " << (*itBegin).c_str();
+    std::cout << (*itBegin).c_str() << " ";
     itBegin++;
   }
